add type DiaryEntry, read the first few lines and output summary
a few rough edges for me here: I don't quiet understand everything
I've done here. but all in all I've refactored the code a bit: now,
it's a two pass approach, first information is gathered in a sequence
of DiaryEntry tuples and then the gathered information is output.

what is missing is some kind of templating mechanism, to factor out
the concern of presenting the information. for that, I need to switch
to nimble, so that I can declare dependencies.